requirements.txt
celery==5.4.0
cloudinary==1.41.0
Django==5.0.4
djangorestframework==3.15.2
django-celery-beat==2.7.0
ipython==8.24.0
pillow==10.4.0
python-dotenv==1.0.1
djoser==2.2.3
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
django-filter==24.3
scikit-learn==1.5.2
pandas==2.2.2
joblib==1.4.2